ui-backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
pillow==10.1.0
werkzeug==3.0.1